exposure creation and dbt_expecations installation
created the file dashboard.yml containing an exposure linked to a dashboard created on Preset platform. Such dashboard is an Executive dashboard that analyses different aspects of mart_full_moon_reviews.sql; Edited packages.yml and added line 4 and 5 to install dbt_expactations. The aim is to use specific tests contained in the package